// src/lib/types.ts
export interface Game {
  id: string;
  name: string;
  description: string;
  thumbnailUrl: string;
  gameUrl: string;
}

export interface GithubContent {
  name: string;
  path: string;
  sha: string;
  size: number;
  url: string;
  html_url: string;
  git_url: string;
  download_url: string | null;
  type: "file" | "dir";
  _links: {
    self: string;
    git: string;
    html: string;
  };
}